Refactor doc and fam lookup tests
